#329 Removed LightBDD.UnitTests.Helpers
