+++ LightBDD.sln
@@ -21,8 +21,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightBDD.Core", "src\LightB
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightBDD.Core.UnitTests", "test\LightBDD.Core.UnitTests\LightBDD.Core.UnitTests.csproj", "{08063D55-3F5E-4CA0-978E-00C6699A3011}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightBDD.UnitTests.Helpers", "test\LightBDD.UnitTests.Helpers\LightBDD.UnitTests.Helpers.csproj", "{43976AB8-60F1-4E2B-855C-0425AC0D053A}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightBDD.AcceptanceTests", "test\LightBDD.AcceptanceTests\LightBDD.AcceptanceTests.csproj", "{17CF2456-692D-4D8D-996B-59BD067E9175}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LightBDD.Framework", "src\LightBDD.Framework\LightBDD.Framework.csproj", "{76150B64-C200-4893-8375-479E377389CC}"
@@ -84,22 +82,6 @@ Global
 		{08063D55-3F5E-4CA0-978E-00C6699A3011}.Release|x64.Build.0 = Release|Any CPU
 		{08063D55-3F5E-4CA0-978E-00C6699A3011}.Release|x86.ActiveCfg = Release|Any CPU
 		{08063D55-3F5E-4CA0-978E-00C6699A3011}.Release|x86.Build.0 = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|ARM.Build.0 = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|x64.Build.0 = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Debug|x86.Build.0 = Debug|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|ARM.ActiveCfg = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|ARM.Build.0 = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|x64.ActiveCfg = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|x64.Build.0 = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|x86.ActiveCfg = Release|Any CPU
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A}.Release|x86.Build.0 = Release|Any CPU
 		{17CF2456-692D-4D8D-996B-59BD067E9175}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{17CF2456-692D-4D8D-996B-59BD067E9175}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{17CF2456-692D-4D8D-996B-59BD067E9175}.Debug|ARM.ActiveCfg = Debug|Any CPU
@@ -203,7 +185,6 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{1404165C-13B2-43E2-A78D-6B271EB62FC8} = {795CD637-D3F6-44AB-A408-AA89B7C18356}
 		{08063D55-3F5E-4CA0-978E-00C6699A3011} = {A7491D1D-7E5B-4021-8105-4098683807EB}
-		{43976AB8-60F1-4E2B-855C-0425AC0D053A} = {A7491D1D-7E5B-4021-8105-4098683807EB}
 		{17CF2456-692D-4D8D-996B-59BD067E9175} = {A7491D1D-7E5B-4021-8105-4098683807EB}
 		{76150B64-C200-4893-8375-479E377389CC} = {795CD637-D3F6-44AB-A408-AA89B7C18356}
 		{89A86E84-BC70-4F54-8022-1FFCAF1D908D} = {A7491D1D-7E5B-4021-8105-4098683807EB}
