+++ test/LightBDD.Core.UnitTests/Execution/ExecutionPipeline_metadata_collection_for_steps_tests.cs
@@ -2,8 +2,6 @@
 using LightBDD.Core.Results;
 using LightBDD.Core.UnitTests.Helpers;
 using LightBDD.ScenarioHelpers;
-using LightBDD.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 
 namespace LightBDD.Core.UnitTests.Execution;
