+++ test/LightBDD.Core.UnitTests/Execution/RunnableScenario_complex_parameter_step_execution_tests.cs
@@ -12,8 +12,6 @@ using LightBDD.Core.UnitTests.Helpers;
 using LightBDD.Framework;
 using LightBDD.Framework.Extensibility;
 using LightBDD.ScenarioHelpers;
-using LightBDD.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using Shouldly;
