+++ test/LightBDD.Core.UnitTests/Execution/RunnableScenario_fixture_management_tests.cs
@@ -6,7 +6,7 @@ using LightBDD.Core.Configuration;
 using LightBDD.Core.Extensibility.Execution;
 using LightBDD.Core.Results;
 using LightBDD.Core.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using Shouldly;
