+++ test/LightBDD.Core.UnitTests/Execution/RunnableScenario_step_asynchronous_execution_tests.cs
@@ -4,7 +4,6 @@ using LightBDD.Core.Extensibility;
 using LightBDD.Core.Results;
 using LightBDD.Core.UnitTests.Helpers;
 using LightBDD.ScenarioHelpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using Shouldly;
