+++ test/LightBDD.Core.UnitTests/Extensibility/CoreMetadataProvider_culture_tests.cs
@@ -1,7 +1,6 @@
 ﻿using LightBDD.Core.Configuration;
 using LightBDD.Core.Extensibility;
 using LightBDD.Core.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using System.Globalization;
 using LightBDD.ScenarioHelpers;
