+++ test/LightBDD.Core.UnitTests/Extensibility/CoreMetadataProvider_tests.cs
@@ -11,7 +11,6 @@ using LightBDD.Core.Metadata;
 using LightBDD.Core.UnitTests.Helpers;
 using LightBDD.Framework;
 using LightBDD.ScenarioHelpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using Shouldly;
