+++ test/LightBDD.Core.UnitTests/Extensibility/RunnableScenarioFactory_tests.cs
@@ -5,7 +5,7 @@ using LightBDD.Core.Extensibility;
 using LightBDD.Core.Extensibility.Execution;
 using LightBDD.Core.Results;
 using LightBDD.Core.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using Moq;
 using NUnit.Framework;
 using Shouldly;
