+++ test/LightBDD.Framework.UnitTests/LightBDD.Framework.UnitTests.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\LightBDD.Framework\LightBDD.Framework.csproj" />
-    <ProjectReference Include="..\LightBDD.UnitTests.Helpers\LightBDD.UnitTests.Helpers.csproj" />
+    <ProjectReference Include="..\LightBDD.ScenarioHelpers\LightBDD.ScenarioHelpers.csproj" />
   </ItemGroup>
 
   <ItemGroup>
