+++ test/LightBDD.Framework.UnitTests/Reporting/FeatureReportExtensions_tests.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using LightBDD.Core.Results;
 using LightBDD.Framework.Reporting;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using NUnit.Framework;
 
 namespace LightBDD.Framework.UnitTests.Reporting
