+++ test/LightBDD.Framework.UnitTests/Reporting/Formatters/HtmlReportFormatter_customization_tests.cs
@@ -8,7 +8,7 @@ using LightBDD.Core.Reporting;
 using LightBDD.Framework.Configuration;
 using LightBDD.Framework.Reporting;
 using LightBDD.Framework.Reporting.Formatters;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using Moq;
 using NUnit.Framework;
