+++ test/LightBDD.Framework.UnitTests/Reporting/Formatters/HtmlReportFormatter_tests.cs
@@ -9,7 +9,7 @@ using LightBDD.Core.Results;
 using LightBDD.Framework.Parameters;
 using LightBDD.Framework.Reporting.Formatters;
 using LightBDD.Framework.UnitTests.Reporting.Formatters.Helpers;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using NUnit.Framework;
 
 namespace LightBDD.Framework.UnitTests.Reporting.Formatters
