+++ test/LightBDD.Framework.UnitTests/Reporting/Formatters/XmlReportFormatter_tests.cs
@@ -6,7 +6,7 @@ using System.Xml.Schema;
 using LightBDD.Core.Results;
 using LightBDD.Framework.Parameters;
 using LightBDD.Framework.Reporting.Formatters;
-using LightBDD.UnitTests.Helpers;
+using LightBDD.ScenarioHelpers;
 using NUnit.Framework;
 
 namespace LightBDD.Framework.UnitTests.Reporting.Formatters
