+++ test/LightBDD.Framework.UnitTests/Scenarios/Extended/Extended_scenario_exception_stack_trace_integration_tests.cs
@@ -4,8 +4,7 @@ using System.Threading.Tasks;
 using LightBDD.Framework.Extensibility;
 using LightBDD.Framework.Scenarios;
 using LightBDD.Framework.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
+using LightBDD.ScenarioHelpers;
 using NUnit.Framework;
 using Shouldly;
