+++ test/LightBDD.Framework.UnitTests/Scenarios/Fluent/Fluent_scenario_runner_execution_tests.cs
@@ -1,6 +1,5 @@
 using LightBDD.Framework.Extensibility;
 using LightBDD.Framework.UnitTests.Scenarios.Extended.Helpers;
-using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using System;
 using System.Threading.Tasks;
