+++ test/LightBDD.ScenarioHelpers/StringExtensions.cs
@@ -1,0 +1,7 @@
+namespace LightBDD.ScenarioHelpers
+{
+    public static class StringExtensions
+    {
+        public static string NormalizeNewLine(this string txt) => txt.Replace("\r", "");
+    }
+}
