--- test/LightBDD.UnitTests.Helpers/LightBDD.UnitTests.Helpers.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <Import Project="..\Common.Tests.props" />
-  <PropertyGroup>
-    <TargetFrameworks>net6;net48</TargetFrameworks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\LightBDD.Framework\LightBDD.Framework.csproj" />
-    <ProjectReference Include="..\LightBDD.ScenarioHelpers\LightBDD.ScenarioHelpers.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="NUnit" Version="3.13.3" />
-    <PackageReference Include="Moq" Version="4.18.4" />
-    <PackageReference Include="RandomTestValues" Version="2.1.1" />
-  </ItemGroup>
-
-</Project>
