--- test/LightBDD.UnitTests.Helpers/StringExtensions.cs
@@ -1,10 +1,0 @@
-namespace LightBDD.UnitTests.Helpers
-{
-    public static class StringExtensions
-    {
-        public static string NormalizeNewLine(this string txt)
-        {
-            return txt.Replace("\r", "");
-        }
-    }
-}
